Hw_07
 I created two fire_noise textures for use with a Niagara emitter fountain  preset. I hooked up the fire-noise to a material in a manner identical to the particle_round_MAT1. I then created fountain emitter preset called campfire  and placed it inside camp_fire_system.
I then attached Camp_fire_particle-MAT1 that to to the camp_fire_System and set gravity slightly positive while increasing size of the particles and decreasing their particle life to get a flame effect. I put the camp_fire_system into the map and moved it to the fire pit I changed the color of the flame to blue and added a light render.
I created a second fountain for sparks added a light render to it and decreased its spawn rate after putting it in the camp_fire_system. 
I then created a smoke_MAT using fire_noise_2 and attached that to a fountain emitter inside of a Smoke_System. I increased the size of its particles and increased the particle life and size to achieve a smoke effect and duplicated the smoke emitter inside the Smoke_System. I then added tilt the direction of the sparks and the smoke to simulate a breeze. 

In summery
 I added a fire, sparks, and smoke