HW_08
I added a ring emitter and a static mesh emitter with a planer mobile texture,
I also created a custom cylinder for the mobile texture,
added a fire effect in a ring inside the mesh cylinder.

I then added a omni directional burst to the trigger that had a green glow and new flame texture

I then added ribbons to the burst emitter 

I added light renders to both the ribbon and the the burst.

I attempted to do something with a beam emitter bu ran out of time